增加面板绑定代码

+++ Assets/Editor/BuildHotfix.cs
@@ -7,9 +7,12 @@ public class BuildHotfix
 {
     static BuildHotfix()
     {
-        File.Copy("Library/ScriptAssemblies/Unity.Hotfix.dll", "Assets/Resources/Text/Unity.Hotfix.dll.bytes", true);
-        File.Copy("Library/ScriptAssemblies/Unity.Hotfix.pdb", "Assets/Resources/Text/Unity.Hotfix.pdb.bytes", true);
-        Debug.Log("复制Hotfix.dll Hotfix.pdb");
-        AssetDatabase.Refresh();
+        if (File.Exists("Library/ScriptAssemblies/Unity.Hotfix.dll") && File.Exists("Library/ScriptAssemblies/Unity.Hotfix.pdb"))
+        {
+            File.Copy("Library/ScriptAssemblies/Unity.Hotfix.dll", "Assets/Resources/Text/Unity.Hotfix.dll.bytes", true);
+            File.Copy("Library/ScriptAssemblies/Unity.Hotfix.pdb", "Assets/Resources/Text/Unity.Hotfix.pdb.bytes", true);
+            Debug.Log("复制Hotfix.dll Hotfix.pdb");
+            AssetDatabase.Refresh();
+        }
     }
 }
